backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-CORS==3.0.10
python-dotenv==0.19.1
# Add any other Python dependencies here